+++ Workspace/DiamondKill/src/com/github/calfur/minecraftserverplugins/diamondkill/disabling/AnvilDisabler.java
@@ -1,4 +1,4 @@
-package com.github.calfur.minecraftserverplugins.diamondkill;
+package com.github.calfur.minecraftserverplugins.diamondkill.disabling;
 
 import java.util.Arrays;
 import java.util.List;
